UGC-4131 | Unbreak extension
In 50e64e5, the UsingData extension was refactored as part of upgrade
work. This sadly broke the $sourcePage check in `getDataFrame`, causing
it to render no data in the using tag and parser function. So, restore
the original, working, condition, and add parser tests for the logic.

Reproducing the issue in parser tests proved a bit nettlesome as the
tests were falsely passing at first. This is because the UsingData
parser hook handler instance uses shared state to store data frame
instances, and this instance is kept alive and reused by all parser
tests, causing data to be shared between tests that would not be shared
between actual page parses. As such, introduce a ParserClearState
handler to reset internal handler state between full parses to make the
tests better reflect the behavior in real parsing. This in turn required
removing the clearState() call and hooks manipulation in getDataFrame(),
as it would also reset our clearState() handler. Given that direct
manipulation of $wgHooks is deprecated as of MediaWiki 1.40 and clearing
the set of handlers for a hook is only allowed in tests going forward,
this is something we'd need to do sooner than later anyways.